Command use in discord, and the beginning of framework capability
The end idea will be to allow for other script to make use of ircBridge to post and recieve commands. Currently the only active command available to IRC is /list which, obviously would list the active players on the server. Due to the IRC connection and matterbridge commands would not check for discord rank.

The additions to framework capability allow for other scripts like serverShutdown by Urm: https://github.com/tes3mp-scripts/ShutdownServer to be able to post when the server it about to shut down. Helpful for large servers and players to know they may wish to wait to log in.